move DownstreamMessages and SubmitShareWithChannelId to utils and make changes accordingly across
